using tenant-scoped role assignments to determine which users a user's user-collection rule(s) should apply to instead of domain
